Crate gauge_sys

This module provides a safe layer for interacting with the legacy Gauge API provided with Microsoft Flight Simulator.

Modules

ffi
Low-level implementation details providing the required FFI bindings for the legacy Gauge API.

Macros

aircraft_variable
Produces a base new aircraft variable

gauge_unit
Produces a new unit type, used for communicating with the gauge API.

indexed_aircraft_variable
Provides access to an aircraft variable that is indexed

named_variable
Constructs bindings for interacting with a named variable

unindexed_aircraft_variable
Provides access to variables that do not require indexing

Traits

AircraftVariable
Identifies an aircraft variable that can be interacted with through the Gauge API.

NamedVariable
A custom named variable that can be interacted with in the Gauge API

Unit
A unit used with the Gauge API.